AI_Ragbot 3/AI_Ragbot/backend/requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
python-multipart==0.0.6
pydantic==2.3.0
chromadb==0.4.18
httpx==0.25.0
pypdf==3.16.0
docx2txt==0.8
sentence-transformers==2.2.2
python-dotenv==1.0.0